+++ changelog.rst
@@ -7,6 +7,7 @@ Last change: 01-DEC-2022 MTS
 -----
 - Cluster info is saved in ``_cluster_centers.hdf5`` files which are created when ``Save cluster centers`` box is ticked
 - Cluster centers contain info about group, mean frame (saved as ``frame``), standard deviation frame, area/volume and convex hull
+- ``gist_rainbow`` is used for rendering properties
 - Bug fixes
 
 0.5.1 - 0.5.4